+++ test/unit_tests/compressorFncsTest.cpp
@@ -14,6 +14,7 @@
 
 // HPWHsim
 #include "HPWH.hh"
+#include "unit-test.hh"
 
 struct CompressorFncsTest : public testing::Test
 {
