+++ test/unit_tests/energyBalanceTest.cpp
@@ -14,6 +14,7 @@
 
 // HPWHsim
 #include "HPWH.hh"
+#include "unit-test.hh"
 
 /*
  * energyBalance tests
