+++ test/unit_tests/unit-test-main.cpp
@@ -4,7 +4,6 @@
 // Vendor
 #include <gtest/gtest.h>
 
-
 int main(int argc, char** argv)
 {
     ::testing::InitGoogleTest(&argc, argv);
